requirements.txt
Django==4.2
environs==9.5.0
django-jazzmin==3.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
django-filter==25.1
Pillow==9.5.0
django-cors-headers==4.2.0
qrcode==8.1
python-telegram-bot==13.15